optimized doubles to floats and threads to 4

+++ ft_eventmousemove.c
@@ -6,7 +6,7 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/05/03 15:16:34 by pholster       #+#    #+#                */
-/*   Updated: 2019/05/24 15:38:53 by pholster      ########   odam.nl         */
+/*   Updated: 2019/06/07 19:04:25 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,8 +18,8 @@ int			ft_eventmousemove(int x, int y, t_ftl *ftl)
 		return (0);
 	if (ftl->locked == FALSE && ft_strequ(ftl->name, "Julia"))
 	{
-		ftl->cx = (((double)x * 3.5) / WINDOW_X) - 2.5;
-		ftl->cy = ((double)y / WINDOW_Y);
+		ftl->cx = (((float)x * 3.5) / WINDOW_X) - 2.5;
+		ftl->cy = ((float)y / WINDOW_Y);
 		ftl->fractol_fnc(ftl);
 	}
 	return (0);
