+++ ft_mandelbrotdraw.c
@@ -6,15 +6,15 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/05/16 12:58:21 by pholster       #+#    #+#                */
-/*   Updated: 2019/06/07 13:34:06 by pholster      ########   odam.nl         */
+/*   Updated: 2019/06/07 19:04:25 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fractol.h"
 
-static int	calccount(t_ftl *ftl, int count, double zx, double zy)
+static int	calccount(t_ftl *ftl, int count, float zx, float zy)
 {
-	double	log_zn;
+	float	log_zn;
 
 	if (count < ftl->detail)
 	{
@@ -25,11 +25,11 @@ static int	calccount(t_ftl *ftl, int count, double zx, double zy)
 	return (count);
 }
 
-static void	calczxy(double *zx, double *zy, double cx, double cy)
+static void	calczxy(float *zx, float *zy, float cx, float cy)
 {
-	double	tempzx;
-	double	tempzy;
-	double	tempx;
+	float	tempzx;
+	float	tempzy;
+	float	tempx;
 
 	tempzx = *zx;
 	tempzy = *zy;
@@ -38,13 +38,13 @@ static void	calczxy(double *zx, double *zy, double cx, double cy)
 	*zx = tempx;
 }
 
-static void	drawx(t_ftl *ftl, int x, int y, double *scale)
+static void	drawx(t_ftl *ftl, int x, int y, float *scale)
 {
 	int			count;
-	double		zx;
-	double		zy;
-	double		cx;
-	double		cy;
+	float		zx;
+	float		zy;
+	float		cx;
+	float		cy;
 
 	cx = x * scale[0] + ftl->posx;
 	cy = y * scale[1] + ftl->posy;
@@ -61,7 +61,7 @@ static void	drawx(t_ftl *ftl, int x, int y, double *scale)
 		ft_putpixel(ftl, x, y, count);
 }
 
-static void	drawy(t_ftl *ftl, atomic_int *renderd, int y, double *scale)
+static void	drawy(t_ftl *ftl, atomic_int *renderd, int y, float *scale)
 {
 	int x;
 
@@ -77,9 +77,9 @@ static void	drawy(t_ftl *ftl, atomic_int *renderd, int y, double *scale)
 void		ft_mandelbrotdraw(t_ftl *ftl)
 {
 	atomic_int	renderd;
-	double		scalex;
-	double		scaley;
-	double		scale[2];
+	float		scalex;
+	float		scaley;
+	float		scale[2];
 	int			y;
 
 	y = 0;
